Test dataset includes UUIDs. Change startDeception to pass the villain ID to Secret.
The dataset needs to include the UUIDs, otherwise there will be null pointer exceptions when they are put into sets. As the UUIDs are strings, for the test data it has been set to the lawyers' anesm.

The declaration for startDeception has been modified to pass the villain's id, as Secret will need this for the cover-up.